Added player to engine, calculates central location on the screen. Firing debug statements for android fling gestures.

+++ app/src/main/java/game/components/PositionComponent.java
@@ -7,10 +7,10 @@ import com.badlogic.ashley.core.Component;
  * Created by Steven on 7/29/2015.
  */
 public class PositionComponent extends Component {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public PositionComponent(int x, int y) {
+    public PositionComponent(float x, float y) {
         this.x = x;
         this.y = y;
     }
